Remove "Sentry Daily Update.xml" scheduled task XML export file.
Best not to include this in a public repo.

Local note for UoY Sysadmins
The specific parameters that we use are in a scheduled task export XML file located in the private repo https://github.com/digital-york/telepen-sentry-scripts-and-conf → `scheduled_tasks_export\Sentry Daily Update.xml` file.
For the most up-to-date copy of this file that schedules the old VBScript script, browse to the following git ref:
https://github.com/digital-york/telepen-sentry-scripts-and-conf/commit/684eca0a852d7c766ae12d96d5f85fe8544ffafc